Refactor rpc server service
Did some refactoring to improve the code extensibility

We want to create more grpc service in the future and this refactoring
is a intermediate step.

N/A

pr-link: Alluxio/alluxio#16774
